+++ src/app/pages/PerpetualPage/components/CompetitionPage/components/Leaderboard/types.ts
@@ -1,0 +1,7 @@
+import { Nullable } from 'types';
+
+export type TableData<T> = {
+  items: T[];
+  userData: Nullable<T>;
+  loaded: boolean;
+};
